# backend/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.2
  title: real-time-river-level-korea-api
  description: '실시한 하천 수위 현황 API 명세 문서'
tags:
  - name: auth
    description: API 접근 인증 관련 기능
  - name: riverlevel
    description: |
      하천수위정보 관련 기능 (token role: "riverlevel")
  - name: crawling
    description: |
      데이터수집 관련 기능 (token role: "crawler")
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    api_key_set_schema:
      type: object
      description: |
        API key + API secret set schema
      required:
        - api_key
        - api_secret
      properties:
        api_key:
          type: string
          description: |
            대문자, 숫자 포함 '-' 제외 28자 조합 문자열(7자 단위로 '-'로 구분)
          example: 1EEA6DC-JAM4DP2-PHVYPBN-V0XCJ9X
        api_secret:
          type: string
          description: |
            대문자, 숫자 포함 '-' 제외 28자 조합 문자열(7자 단위로 '-'로 구분)
          example: 1EEA6DC-JAM4DP2-PHVYPBN-V0XCJ9X
    JWT_token_set_schema:
      type: object
      description: |
        Access JWT Token과 Refresh JWT Token 반환 형식
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          description: |
            API 이용 접근 용 JWT Token  

            발급 후 아래와 같이 HTTP header Authorization 항목에 Bearer token으로 포함하세요.  

            HTTP header : {  
              Authorization: Bearer <access_token>  
            }
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refresh_token:
          type: string
          description: |
            Access Token 재발급 용 JWT Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c

    sido_param:
      type: string
      description: |
        "시도" 파라미터
      enum:
        [
          서울특별시,
          부산광역시,
          대구광역시,
          인천광역시,
          광주광역시,
          대전광역시,
          울산광역시,
          세종특별자치시,
          경기도,
          강원도,
          충청북도,
          충청남도,
          전라북도,
          전라남도,
          경상북도,
          경상남도,
          제주특별자치도,
        ]
      example: 서울특별시

    sigungu_param:
      type: string
      description: |
        "시군구" 파라미터 ("시도" 파라미터에 따라 달라짐.)

    river_level_response:
      type: object
      description: |
        하천수위정보 반환 형식
      required:
        - sido_name
        - sigungu_length
        - sigungu
      properties:
        sido_name:
          type: string
          description: |
            요청한 "시도"명
        sigungu_length:
          type: integer
          description: |
            요청한 "시도" 하위의 "시군구" 개수
          format: int32
        sigungu:
          type: object
          description: |
            요청한 "시도" 하위의 "시군구" 모음
          additionalProperties:
            type: object
            description: |
              "시군구" 객체
            required:
              - sigungu_name
              - average_river_level_ratio
              - rivergauge_length
              - rivergauge
            properties:
              sigungu_name:
                type: string
                description: |
                  "시군구"명
              average_river_level_ratio:
                type: number
                description: |
                  "시군구"의 평균하천수위비율
                format: double
              rivergauge_length:
                type: integer
                description: |
                  "시군구" 내의 하천수위 관측계 개수
                format: int32
              rivergauge:
                type: object
                description: |
                  "시군구" 내의 하천수위 관측계 모음
                additionalProperties:
                  type: object
                  description: |
                    "시군구" 내의 하천수위 관측계 객체
                  required:
                    - obscd
                    - obsnm
                    - addr
                    - lon
                    - lat
                    - data
                  properties:
                    obscd:
                      type: string
                      description: 관측계 코드
                    obsnm:
                      type: string
                      description: 관측계 이름
                    addr:
                      type: string
                      description: 관측계 주소
                    lon:
                      type: string
                      description: |
                        관측계 경도 좌표 (형식: {도}-{분}-{초})
                    lat:
                      type: string
                      description: |
                        관측계 위도 좌효 (형식: {도}-{분}-{초})
                    planflood_level:
                      type: number
                      format: double
                      description: 계획홍수위(m)
                    gdt:
                      type: number
                      format: double
                      description: 영점표고 (m)
                    data:
                      type: array
                      description: 하천수위 관측계의 하천수위관측 데이터 배열
                      items:
                        type: object
                        description: 하천수위 관측계의 하천수위관측 데이터
                        required:
                          - timestamp
                          - current_level
                        properties:
                          timestamp:
                            type: string
                            description: 관측일시
                            format: date-time
                          current_level:
                            type: number
                            description: 현재하천수위(m)
                            format: double
                          river_level_ratio:
                            type: number
                            description: 하천수위비율(현채하천수위 ÷ 계획홍수위, 0 ~ 1 사이 범위의 실수)
                            format: double

paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: 신규 API key / API secret 발급
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - owner
                - admin_key
              properties:
                role:
                  type: string
                  description: |
                    API 사용허가영역 (","를 통해 구문하여 복수 개의 사용허가영역 적용 가능)
                    |  사용허가영역 |                  설명                      |
                    | ---------- | ----------------------------------------- |
                    | riverlevel | 하천수위정보 API를 이용할 수 있도록 허용된 분류입니다. |
                    |   crawler  | 하천수위정보를 등록할 수 있도록 허용된 분류입니다.     |
                  enum: [riverlevel, crawler]
                  example: riverlevel
                owner:
                  type: string
                  description: |
                    API key set 소유자명  
                    대문자, 소문자, 숫자 내에서 3자 이상 20자 이하 조합
                  example: honggildong
                admin_key:
                  type: string
                  description: |
                    관리자 비밀키
                  example: ADMIN_SECRET_KEY
                description:
                  type: string
                  description: 해당 API key set의 설명
                  example: honggildong's API key set

      responses:
        '200':
          description: 요청 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_key_set_schema'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 실패
        '5XX':
          description: 내부 오류

  /auth/signin:
    post:
      tags:
        - auth
      summary: API key / API secret를 통한 Access Token과 Refresh Token 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api_key_set_schema'
      responses:
        '200':
          description: 인증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT_token_set_schema'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 실패
        '5XX':
          description: 내부 오류

  /auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh Token을 통한 Access Token과 Refresh Token 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
      responses:
        '200':
          description: 인증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT_token_set_schema'
        '400':
          description: 잘못된 요청
        '5XX':
          description: 내부 오류

  /riverlevel/{sido}:
    get:
      tags:
        - riverlevel
      summary: 요청한 "시도" 내 모든 "시군구"의 관측계 별 하천수위 정보
      description: |
        파라미터 **"sido"**를 통해 모든 "시군구"의 관측계 별 하천수위 정보를 가져올 "시도"를 지정할 수 있습니다.  
        **각 관측계의 "planflood_level", "gdt", "river_level_ratio"는 null 값일 수 있습니다.**
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sido
          required: true
          schema:
            $ref: '#/components/schemas/sido_param'
          description: |
            요청할 "시도"의 한글표기(접미사 표현까지 기입)  
            ex) 서울특별시, 세종특별자치시, 경상북도, 제주특별자치도
      responses:
        '200':
          description: |
            요청 성공  
            각 "시군구" 안에는 관측소 별 하천수위 정보가 포함됩니다.  
            단, 각 관측소 별로 단 한 개의 하천수위 정보만 포함됩니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/river_level_response'
              example:
                sido_name: 서울특별시
                sigungu_length: 2
                sigungu:
                  성동구:
                    sigungu_name: 성동구
                    average_river_level_ratio: 0.38856304
                    riverguage_length: 2
                    riverguage:
                      성동교:
                        obscd: seoul_901
                        obsnm: 성동교
                        addr: 서울특별시 성동구 행당동 성동교
                        lon: 127-2-37
                        lat: 37-33-9
                        planflood_level: 17.05
                        gdt: null
                        data:
                          - timestamp: '2022-08-17T13:30:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                      마장2교:
                        obscd: seoul_902
                        obsnm: 마장2교
                        addr: 서울시 성동구 마장동 마장2교
                        lon: 127-2-48
                        lat: 37-34-7
                        planflood_level: 17.05
                        gdt: null
                        data:
                          - timestamp: '2022-08-17T13:30:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
                  마포구:
                    sigungu_name: 마포구
                    average_river_level_ratio: 0.76339869
                    riverguage_length: 1
                    riverguage:
                      성산2교:
                        obscd: seoul_1501
                        obsnm: 성산2교
                        addr: 서울특별시 마포구 성산동 성산2교
                        lon: 37-33-56
                        lat: 126-54-41
                        planflood_level: 15.3
                        gdt: null
                        data:
                          - timestamp: '2022-08-17T13:30:15Z'
                            current_level: 11.68
                            river_level_ratio: 0.76339869
        '400':
          description: 잘못된 요청
        '401':
          description: 잘못되거나 만료된 인증 정보
        '403':
          description: 허용되지 않은 권한
        '5XX':
          description: 내부 오류

  /riverlevel/{sido}/{sigungu}:
    get:
      tags:
        - riverlevel
      summary: 요청한 "시도" 내 특정 "시군구"의 관측계 별 하천수위 정보
      description: |
        파라미터 **"sido"**와 **"sigungu"**를 통해 관측계 별 하천수위 정보를 가져올 "시도"와 "시군구"를 지정할 수 있습니다.  
        **각 관측계의 "planflood_level", "gdt", "river_level_ratio"는 null 값일 수 있습니다.**
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: sido
          required: true
          schema:
            $ref: '#/components/schemas/sido_param'
          description: |
            요청할 "시도"의 한글표기(접미사 표현까지 기입)  
            ex) 서울특별시, 세종특별자치시, 경상북도, 제주특별자치도
        - in: path
          name: sigungu
          required: true
          schema:
            $ref: '#/components/schemas/sigungu_param'
          example: '성동구'
          description: |
            요청할 "시도" 하위 특정 "시군구"의 한글표기(접미사 표현까지 기입)
            (**주의**: "시도" 하위에 존재하지 않는 "시군구"를 요청할 경우 400 반환)
      responses:
        '200':
          description: |
            요청 성공  
            요청한 "시도" 하위 특정 "시군구" 하천들의 수위정보만 반환한다.  
            이때 각 하천 별 수위정보들은 시간대 별로 여러개가 반환될 수 있다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/river_level_response'
              example:
                sido_name: 서울특별시
                average_river_level_ratio: 0.9
                sigungu_length: 1
                sigungu:
                  성동구:
                    sigungu_name: 성동구
                    average_river_level_ratio: 0.38856304
                    riverguage_length: 2
                    riverguage:
                      성동교:
                        obscd: seoul_901
                        obsnm: 성동교
                        addr: 서울특별시 성동구 행당동 성동교
                        lon: 127-2-37
                        lat: 37-33-9
                        planflood_level: 17.05
                        gdt: null
                        data:
                          - timestamp: '2022-08-17T13:10:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                          - timestamp: '2022-08-17T13:15:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                          - timestamp: '2022-08-17T13:20:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                          - timestamp: '2022-08-17T13:25:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                          - timestamp: '2022-08-17T13:30:15Z'
                            current_level: 4.96
                            river_level_ratio: 0.29032258
                      마장2교:
                        obscd: seoul_902
                        obsnm: 마장2교
                        addr: 서울시 성동구 마장동 마장2교
                        lon: 127-2-48
                        lat: 37-34-7
                        planflood_level: 17.05
                        gdt: null
                        data:
                          - timestamp: '2022-08-17T13:10:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
                          - timestamp: '2022-08-17T13:15:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
                          - timestamp: '2022-08-17T13:20:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
                          - timestamp: '2022-08-17T13:25:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
                          - timestamp: '2022-08-17T13:30:15Z'
                            current_level: 8.3
                            river_level_ratio: 0.48680351
        '400':
          description: 잘못된 요청
        '401':
          description: 잘못되거나 만료된 인증 정보
        '403':
          description: 허용되지 않은 권한
        '5XX':
          description: 내부 오류

  /crawling/riverlevel:
    post:
      tags:
        - crawling
      summary: 하천 수위 데이터를 추가
      description: |
        등록된 수위 관측계에 대한 하천 수위 데이터를 추가할 수 있습니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                data:
                  - obscd: seoul_1501
                    water_level: 1.5
                    obsdate: '2022-08-17T13:30:15Z'
              required:
                - data
              properties:
                data:
                  type: array
                  description: 수위 관측계 관측데이터 모음
                  items:
                    type: object
                    description: 수위 관측계의 수위관측데이터
                    required:
                      - obscd
                      - water_level
                      - obsdate
                    properties:
                      obscd:
                        type: string
                        description: 수위 관측계 코드
                      water_level:
                        type: number
                        description: 현재하천수위(m)
                        format: double
                      obsdate:
                        type: string
                        description: 관측일시
                        format: date-time
      responses:
        '200':
          description: 요청 성공
        '400':
          description: 잘못된 요청
        '401':
          description: 잘못되거나 만료된 인증 정보
        '403':
          description: 허용되지 않은 권한
        '5XX':
          description: 내부 오류

  /crawling/rivergauge:
    post:
      tags:
        - crawling
      summary: 수위 관측계를 신규 등록
      description: |
        새로운 하천 수위 관측계를 신규 등록할 수 있습니다. 단, 기존에 등록된 관측계 코드로 관측계의 재등록이 불가합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              example:
                data:
                  - obscd: seoul_1501
                    obsnm: 성산2교
                    mngorg: 환경부
                    addr: 서울특별시 마포구 성산동 성산2교
                    flood_warning: N
                    lon: 37-33-56
                    lat: 126-54-41
                    planflood_level: 15.3
                    gdt: null
              properties:
                data:
                  type: array
                  description: 수위 관측계 정보 모음
                  items:
                    type: object
                    description: 수위 관측계 정보
                    required:
                      - obscd
                      - obsnm
                      - mngorg
                      - flood_warning
                      - planflood_level
                      - gdt
                    properties:
                      obscd:
                        type: string
                        description: 수위 관측계 코드
                      obsnm:
                        type: string
                        description: 수위 관측계 이름
                      mngorg:
                        type: string
                        description: 수위 관측계 관할 기관명
                      flood_warning:
                        type: string
                        enum: [Y, N]
                        description: 홍수특보발령여부
                      addr:
                        type: string
                        description: 수위 관측계 주소
                      lon:
                        type: string
                        description: 수위 관측계 경도 좌표 ("-"로 구분하여 {도}-{분}-{초}로 기입)
                      lat:
                        type: string
                        description: 수위 관측계 위도 좌효 ("-"로 구분하여 {도}-{분}-{초}로 기입)
                      gdt:
                        type: number
                        description: 영점표고
                        format: double
                      planflood_level:
                        type: number
                        description: 계획홍수위(m)
                        format: double
      responses:
        '200':
          description: 요청 성공
        '400':
          description: 잘못된 요청
        '401':
          description: 잘못되거나 만료된 인증 정보
        '403':
          description: 허용되지 않은 권한
        '5XX':
          description: 내부 오류
    patch:
      tags:
        - crawling
      summary: 수위 관측계의 등록 정보를 수정
      description: |
        등록된 수위 관측계의 등록정보를 수정할 수 있습니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                data:
                  - obscd: seoul_1501
                    obsnm: 성산2교
                    mngorg: 환경부
                    addr: 서울특별시 마포구 성산동 성산2교
                    flood_warning: N
                    lon: 37-33-56
                    lat: 126-54-41
                    planflood_level: 15.3
                    gdt: null
              required:
                - data
              properties:
                data:
                  type: array
                  description: 수위 관측계 정보 모음
                  items:
                    type: object
                    description: 수위 관측계 정보
                    required:
                      - obscd
                    properties:
                      obscd:
                        type: string
                        description: 수위 관측계 코드
                      obsnm:
                        type: string
                        description: 수위 관측계 이름
                      mngorg:
                        type: string
                        description: 수위 관측계 관할 기관명
                      flood_warning:
                        type: string
                        enum: [Y, N]
                        description: 홍수특보발령여부
                      addr:
                        type: string
                        description: 수위 관측계 주소
                      lon:
                        type: string
                        description: 수위 관측계 경도 좌표 ("-"로 구분하여 {도}-{분}-{초}로 기입)
                      lat:
                        type: string
                        description: 수위 관측계 위도 좌효 ("-"로 구분하여 {도}-{분}-{초}로 기입)
                      gdt:
                        type: number
                        description: 영점표고
                        format: double
                      planflood_level:
                        type: number
                        description: 계획홍수위(m)
                        format: double
      responses:
        '200':
          description: 요청 성공
        '400':
          description: 잘못된 요청
        '401':
          description: 잘못되거나 만료된 인증 정보
        '403':
          description: 허용되지 않은 권한
        '5XX':
          description: 내부 오류

servers:
  - description: localhost
    url: http://localhost:3000